+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # News
 
+## v0.1.1 - 2025-07-31
+
+- T1, T2 Noise, visualizer added at test/view_T1T2_insertion.jl
+- Measurement error
+- Compat requires Quantikz, and BPGates 1.2.2
+
 ## v0.1.0 - 2025-06-13
 
 - Simulation capabilities for commond gates.